fix: address copilot review comments on PR #34
- Restore screen reader accessibility for serving state by adding
  back aria-describedby and visually-hidden status node
- Extract duplicated resolveCssColor test helper into shared
  test/utils/css.ts module
