+++ src/components/ActiveMatchScreen/TeamPanel/TeamPanel.module.css
@@ -89,6 +89,18 @@
   color: var(--team-serving-score-color);
 }
 
+.srOnly {
+  position: absolute;
+  inline-size: 1px;
+  block-size: 1px;
+  padding: 0;
+  margin: -1px;
+  overflow: hidden;
+  clip-path: inset(50%);
+  white-space: nowrap;
+  border: 0;
+}
+
 @media (width <= 767px) {
   .panel {
     min-height: 280px;
